refactor(core): decouple T1 button handling from touch
[no changelog]
