+++ core/SConscript.firmware
@@ -347,7 +347,6 @@ SOURCE_TREZORHAL = [
     'embed/trezorhal/rng.c',
     'embed/trezorhal/stm32.c',
     'embed/trezorhal/systick.c',
-    'embed/trezorhal/touch.c',
     'embed/trezorhal/usb.c',
     'embed/trezorhal/usbd_conf.c',
     'embed/trezorhal/usbd_core.c',
@@ -360,6 +359,11 @@ if TREZOR_MODEL == 'T':
     SOURCE_TREZORHAL += [
         'embed/trezorhal/sbu.c',
         'embed/trezorhal/sdcard.c',
+        'embed/trezorhal/touch.c',
+    ]
+elif TREZOR_MODEL == '1':
+    SOURCE_TREZORHAL += [
+        'embed/trezorhal/button.c',
     ]
 
 CPPDEFINES_MOD += ['USE_SVC_SHUTDOWN']
