+++ core/embed/extmod/modtrezorio/modtrezorio.c
@@ -27,6 +27,7 @@
 
 #include <unistd.h>
 
+#include "button.h"
 #include "touch.h"
 #include "usb.h"
 
@@ -59,6 +60,12 @@
 /// TOUCH_MOVE: int  # event id of touch move event
 /// TOUCH_END: int  # event id of touch end event
 
+/// BUTTON: int  # interface id of button events
+/// BUTTON_PRESSED: int  # button down event
+/// BUTTON_RELEASED: int  # button up event
+/// BUTTON_LEFT: int  # button number of left button
+/// BUTTON_RIGHT: int  # button number of right button
+
 /// WireInterface = Union[HID, WebUSB]
 
 /// if False:
@@ -71,6 +78,19 @@ STATIC const mp_rom_map_elem_t mp_module_trezorio_globals_table[] = {
     {MP_ROM_QSTR(MP_QSTR_fatfs), MP_ROM_PTR(&mod_trezorio_fatfs_module)},
     {MP_ROM_QSTR(MP_QSTR_SBU), MP_ROM_PTR(&mod_trezorio_SBU_type)},
     {MP_ROM_QSTR(MP_QSTR_sdcard), MP_ROM_PTR(&mod_trezorio_sdcard_module)},
+
+    {MP_ROM_QSTR(MP_QSTR_TOUCH), MP_ROM_INT(TOUCH_IFACE)},
+    {MP_ROM_QSTR(MP_QSTR_TOUCH_START), MP_ROM_INT((TOUCH_START >> 24) & 0xFFU)},
+    {MP_ROM_QSTR(MP_QSTR_TOUCH_MOVE), MP_ROM_INT((TOUCH_MOVE >> 24) & 0xFFU)},
+    {MP_ROM_QSTR(MP_QSTR_TOUCH_END), MP_ROM_INT((TOUCH_END >> 24) & 0xFFU)},
+#elif TREZOR_MODEL == 1
+    {MP_ROM_QSTR(MP_QSTR_BUTTON), MP_ROM_INT(BUTTON_IFACE)},
+    {MP_ROM_QSTR(MP_QSTR_BUTTON_PRESSED),
+     MP_ROM_INT((BTN_EVT_DOWN >> 24) & 0x3U)},
+    {MP_ROM_QSTR(MP_QSTR_BUTTON_RELEASED),
+     MP_ROM_INT((BTN_EVT_UP >> 24) & 0x3U)},
+    {MP_ROM_QSTR(MP_QSTR_BUTTON_LEFT), MP_ROM_INT(BTN_LEFT)},
+    {MP_ROM_QSTR(MP_QSTR_BUTTON_RIGHT), MP_ROM_INT(BTN_RIGHT)},
 #endif
 
     {MP_ROM_QSTR(MP_QSTR_FlashOTP), MP_ROM_PTR(&mod_trezorio_FlashOTP_type)},
@@ -83,11 +103,6 @@ STATIC const mp_rom_map_elem_t mp_module_trezorio_globals_table[] = {
     {MP_ROM_QSTR(MP_QSTR_poll), MP_ROM_PTR(&mod_trezorio_poll_obj)},
     {MP_ROM_QSTR(MP_QSTR_POLL_READ), MP_ROM_INT(POLL_READ)},
     {MP_ROM_QSTR(MP_QSTR_POLL_WRITE), MP_ROM_INT(POLL_WRITE)},
-
-    {MP_ROM_QSTR(MP_QSTR_TOUCH), MP_ROM_INT(TOUCH_IFACE)},
-    {MP_ROM_QSTR(MP_QSTR_TOUCH_START), MP_ROM_INT((TOUCH_START >> 24) & 0xFFU)},
-    {MP_ROM_QSTR(MP_QSTR_TOUCH_MOVE), MP_ROM_INT((TOUCH_MOVE >> 24) & 0xFFU)},
-    {MP_ROM_QSTR(MP_QSTR_TOUCH_END), MP_ROM_INT((TOUCH_END >> 24) & 0xFFU)},
 };
 
 STATIC MP_DEFINE_CONST_DICT(mp_module_trezorio_globals,
