+++ core/embed/trezorhal/button.h
@@ -1,0 +1,34 @@
+/*
+ * This file is part of the Trezor project, https://trezor.io/
+ *
+ * Copyright (c) SatoshiLabs
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef TREZORHAL_BUTTON_H
+#define TREZORHAL_BUTTON_H
+
+#include <stdint.h>
+
+#define BTN_EVT_DOWN (1U << 24)
+#define BTN_EVT_UP (1U << 25)
+
+#define BTN_LEFT 0
+#define BTN_RIGHT 1
+
+void button_init(void);
+uint32_t button_read(void);
+
+#endif
