+++ core/embed/unix/button.h
@@ -1,0 +1,1 @@
+../trezorhal/button.h
